Cleaned up C++ test names.
